Optimize build (#3227)
* Exclude NeurophTest

* Fix build
