+++ flyway/spring-flyway/src/test/java/com/baeldung/springflyway/CustomerRepositoryInitialMigrationIntegrationTest.java
@@ -6,20 +6,23 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.test.context.junit4.SpringRunner;
+
+import static org.junit.Assert.assertNotNull;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class CustomerRepositoryNotNullConstraintMigrationTest {
+public class CustomerRepositoryInitialMigrationIntegrationTest {
 
     @Autowired CustomerRepository customerRepository;
 
-    @Test(expected = DataIntegrityViolationException.class)
-    public void givenTheNotNullConstraintMigrations_whenInsertingACustomerWithNullEmail_thenThrowException() {
-        customerRepository.save(Customer
+    @Test
+    public void givenSchemaCreationMigration_whenTryingToCreateACustomer_thenSuccess() {
+        Customer customer = customerRepository.save(Customer
           .builder()
+          .email("customer@email.com")
           .build());
+        assertNotNull(customer.getId());
     }
 
 }
