+++ pom.xml
@@ -142,7 +142,6 @@
         <module>spark-java</module>
       <!--  <module>spring-5</module>-->
         <module>spring-5-mvc</module>
-        <module>spring-acl</module>
         <module>spring-activiti</module>
         <module>spring-akka</module>
         <module>spring-amqp</module>
@@ -152,7 +151,7 @@
         <module>spring-batch</module>
         <module>spring-bom</module>
         <module>spring-boot</module>
-	<module>spring-boot-keycloak</module>
+	    <module>spring-boot-keycloak</module>
         <module>spring-boot-bootstrap</module>
         <module>spring-cloud-data-flow</module>
         <module>spring-cloud</module>
