--- spring-acl/src/main/java/org/baeldung/acl/config/AclMethodSecurityConfiguration.java
@@ -1,21 +1,0 @@
-package org.baeldung.acl.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.access.expression.method.MethodSecurityExpressionHandler;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.method.configuration.GlobalMethodSecurityConfiguration;
-
-@Configuration
-@EnableGlobalMethodSecurity(prePostEnabled = true, securedEnabled = true)
-public class AclMethodSecurityConfiguration extends GlobalMethodSecurityConfiguration {
-
-    @Autowired
-    MethodSecurityExpressionHandler defaultMethodSecurityExpressionHandler;
-
-    @Override
-    protected MethodSecurityExpressionHandler createExpressionHandler() {
-        return defaultMethodSecurityExpressionHandler;
-    }
-
-}
