--- spring-acl/src/main/java/org/baeldung/acl/persistence/entity/NoticeMessage.java
@@ -1,29 +1,0 @@
-package org.baeldung.acl.persistence.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name="system_message")
-public class NoticeMessage {
-
-    @Id
-    @Column
-    private Integer id;
-    @Column
-    private String content;
-    public Integer getId() {
-        return id;
-    }
-    public void setId(Integer id) {
-        this.id = id;
-    }
-    public String getContent() {
-        return content;
-    }
-    public void setContent(String content) {
-        this.content = content;
-    }
-}
