--- spring-acl/src/main/resources/org.baeldung.acl.datasource.properties
@@ -1,12 +1,0 @@
-spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_ON_EXIT=FALSE
-spring.datasource.username=sa
-spring.datasource.password=
-spring.datasource.driverClassName=org.h2.Driver
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
-
-spring.h2.console.path=/myconsole
-spring.h2.console.enabled=true
-spring.datasource.initialize=true
-spring.datasource.schema=classpath:acl-schema.sql
-spring.datasource.data=classpath:acl-data.sql
